Implement payment flow and UI improvements
Added image assets for UI: cars, cat, graphic, paduti, product, etc.

Updated React components: Pricing, TryNow, Testimonials, Header, Footer, Menubar, BgSlider, BgRemovalStep.

Implemented backend logic for payment: OrderController.java, SecurityConfig.java.

Updated .vscode/settings.json with Java compile null analysis mode.

Minor style adjustments in index.css.

Total: 36 changed files.

+++ removebg/src/main/java/in/paduti/removebg/entity/OrderEntity.java
@@ -1,6 +1,6 @@
 package in.paduti.removebg.entity;
 
-import java.security.Timestamp;
+import java.sql.Timestamp;
 
 import org.hibernate.annotations.CreationTimestamp;
 
@@ -30,7 +30,7 @@ public class OrderEntity {
     private String orderId;
     private String clerkId;
     private String plan;
-    private Double amount;
+    private Long amount;
     private Integer credits;
     private Boolean payment;
     @CreationTimestamp
